Moving toward getting codebase to a state where the datastore has a clearly defined boundary.
The codebase had many “Store” interfaces defined which represented an abstract interface to the datastore. I made the use of these datastore interfaces more consistent in terms of use and naming. I tried to distinguish between the database implementation vs Store abstraction in naming.

When using MySQL as the datastore, BoltDB was still used for a few entities. I made changes so this cannot happen - only the selected datastore implementation is initialized.

I encountered 3 entities not implemented in MySQL : User, Blueprint and SCEP Challenge. I called them out with a TODO comment and a log message. Endpoints related to entities not implemented in MySQL do not response and will cause a 404 not found response code.

I introduced a new enum field in server.Server struct named “Datastore” that indicates the datastore implementation. This is used instead of checking if the MySQL database is not equal to nil (c.MysqlDB != nil) to determine the datastore.